requirements.txt
aiogram==3.4.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
asyncio
